chore(refactor): changes to folders and configs
