+++ server/routes/club/Club.js
@@ -1,10 +1,10 @@
 //* All routes related to club's LOGIN AND REGISTER
 
 const express = require("express");
-const Club = require("../models/ClubsSchema");
+const Club = require("../../schema/club/ClubSchema");
 const router = express.Router();
 const bcrypt = require("bcryptjs");
-const Events = require("../models/EventsSchema");
+const Events = require("../../schema/club/EventSchema");
 var jwt = require("jsonwebtoken");
 
 //* Route 1  - Club Registration
