+++ server/routes/shop/Products.js
@@ -1,7 +1,7 @@
 //* All routes related to Product's
 
 const express = require("express");
-const Products = require("../models/ProductSchema");
+const Products = require("../../schema/shop/ProductSchema");
 const router = express.Router();
 
 //* Route 1  -  Adding Products
