+++ server/routes/user/User.js
@@ -1,13 +1,12 @@
 //* All routes related to user's LOGIN AND REGISTER
 
 const express = require("express");
-const User = require("../models/UserSchema");
+const User = require("../../schema/user/UserSchema");
 const router = express.Router();
 const bcrypt = require("bcryptjs");
 var jwt = require("jsonwebtoken");
-const ReportProblem = require("../models/ReportProblemSchema");
-const ContactUs = require("../models/ContactUsSchema");
-const passport = require("passport");
+const ReportProblem = require("../../schema/user/ReportProblemSchema");
+const ContactUs = require("../../schema/user/ContactUsSchema");
 
 //* Route 1  - User Registration
 router.post("/register", async (req, res) => {
@@ -113,7 +112,6 @@ router.post("/login", async (req, res) => {
 //* Route 3  - Report a Problem
 router.post("/userreport", async (req, res) => {
   try {
-    //fetch previous report from the same user
     const currentHour = new Date().getMinutes();
     const previousReports = await ReportProblem.find({
       email: req.body.email,
@@ -123,7 +121,7 @@ router.post("/userreport", async (req, res) => {
       let hourOfThisReport = new Date(
         previousReports[i].createdAt,
       ).getMinutes();
-      //check if the user created a report in the last 2 hours
+
       if (hourOfThisReport >= currentHour - 120) {
         return res.json({
           success: false,
@@ -131,7 +129,7 @@ router.post("/userreport", async (req, res) => {
         });
       }
     }
-    //else begin to insert the request in database
+
     const data = req.body;
 
     const ReportData = ReportProblem({
@@ -141,7 +139,6 @@ router.post("/userreport", async (req, res) => {
       reportmessage: data.reportmessage,
     });
 
-    //saving the data to mongodb
     await ReportData.save();
     res.status(200).json({ success: true });
   } catch (e) {
@@ -152,9 +149,8 @@ router.post("/userreport", async (req, res) => {
 //* Route 4  - Contact Us
 router.post("/contactus", async (req, res) => {
   try {
-    //insert the Sender's Data in database
     const data = req.body;
-    const email = data.email; // Primary Key
+    const email = data.email;
 
     const SenderData = {
       firstname: data.firstName,
@@ -175,66 +171,6 @@ router.post("/contactus", async (req, res) => {
   } catch (e) {
     res.status(500).json({ message: "Internal Server Error" });
   }
-});
-
-//* Route 5  - google authentication
-router.get("/google", (req, res) => {
-  const googleAuthURL = "https://accounts.google.com/o/oauth2/v2/auth";
-
-  const params = new URLSearchParams({
-    response_type: "code",
-    redirect_uri: process.env.CALLBACK_URL, // Replace with your redirect URI
-    scope: "profile email ",
-    client_id: process.env.CLIENT_ID, // Replace with your client ID
-  });
-
-  const redirectURL = `${googleAuthURL}?${params}`;
-  res.redirect(redirectURL);
-  // res.json({ url: redirectURL });
-});
-
-//* Route 6  - google authentication callback
-router.get(
-  "/google/callback",
-  passport.authenticate("google", {
-    successRedirect: process.env.successURL,
-    failureRedirect: "/login/failed",
-  }),
-);
-
-//* Route 7  - google authentication failed
-router.get("/login/failed", (req, res) => {
-  res
-    .status(401)
-    .json({ error: true, message: "User failed to authenticate." });
-});
-
-//* Route 8  - google authentication success
-router.get("/login/success", (req, res) => {
-  if (req.user) {
-    const data = { User: { id: req.user.email } };
-
-    res.status(200).json({
-      success: true,
-      error: false,
-      message: "Successfully Loged In",
-      user: req.user,
-      accessToken: jwt.sign(data, process.env.JWT_SECRET),
-    });
-  } else {
-    res.status(403).json({ error: true, message: "Not Authorized" });
-  }
-});
-
-//* Route 9  - google authentication logout
-router.get("/logout", (req, res) => {
-  req.logout(function (err) {
-    if (err) {
-      console.log(err);
-      return;
-    }
-    res.json({ success: true });
-  });
 });
 
 module.exports = router;
